text saving is having issue, need to study how to lock db..?

+++ src/main/java/nio_client/domain/Text.java
@@ -16,7 +16,6 @@ public class Text
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-
     private int textNum;
 
     @ManyToOne
@@ -25,7 +24,7 @@ public class Text
 
     private String sender;
     private String text;
-    private long notReadNum;
+    private int notReadNum;
     private String time;
